rpi/requirements.txt
paho-mqtt
gpiozero
picamera2
PyYAML
requests
